@@ -1,5 +1,12 @@
+require "application_responder"
+
 class ApplicationController < ActionController::Base
-  protect_from_forgery
+  self.responder = ApplicationResponder
+  respond_to :html
+
+  # Prevent CSRF attacks by raising an exception.
+  # For APIs, you may want to use :null_session instead.
+  protect_from_forgery with: :exception
 
   before_filter :set_locale
 
